feat!: types used by API calls moved to separate namespaces
