--- src/api/v2/detail.rs
@@ -1,5 +1,0 @@
-#[derive(Debug)]
-pub struct TrackFiltersBuilderError;
-
-#[derive(Debug)]
-pub struct ExerciseFiltersBuilderError;
