+++ src/core.rs
@@ -32,7 +32,7 @@ impl Credentials {
 }
 
 /// Result type used by the [mini_exercism](crate) library when an error can occur.
-pub type Result<T> = std::result::Result<T, Error>;
+pub type Result<T> = core::result::Result<T, Error>;
 
 /// Error type used by the [mini_exercism](crate) library.
 #[derive(Debug, Error)]
