+++ tests/api_integration/v1.rs
@@ -1,7 +1,7 @@
 mod get_solution {
     use assert_matches::assert_matches;
     use mini_exercism::api;
-    use mini_exercism::core::Error;
+    use mini_exercism::Error;
     use reqwest::StatusCode;
 
     #[tokio::test]
@@ -20,7 +20,7 @@ mod get_solution {
 mod get_latest_solution {
     use assert_matches::assert_matches;
     use mini_exercism::api;
-    use mini_exercism::core::Error;
+    use mini_exercism::Error;
     use reqwest::StatusCode;
 
     #[tokio::test]
@@ -38,7 +38,7 @@ mod get_file {
     use assert_matches::assert_matches;
     use futures::StreamExt;
     use mini_exercism::api;
-    use mini_exercism::core::Error;
+    use mini_exercism::Error;
     use reqwest::StatusCode;
 
     #[tokio::test]
@@ -58,7 +58,7 @@ mod get_file {
 mod get_track {
     use assert_matches::assert_matches;
     use mini_exercism::api;
-    use mini_exercism::core::Error;
+    use mini_exercism::Error;
     use reqwest::StatusCode;
 
     #[tokio::test]
