+++ tests/v2/solution.rs
@@ -1,0 +1,189 @@
+mod solution_test {
+    mod deserialize {
+        use mini_exercism::api::v2::solution::MentoringStatus::Finished;
+        use mini_exercism::api::v2::solution::Status::Published;
+        use mini_exercism::api::v2::solution::TestsStatus::Passed;
+        use mini_exercism::api::v2::solution::{
+            Exercise, MentoringStatus, Solution, Status, TestsStatus, Track,
+        };
+
+        #[test]
+        fn test_all() {
+            let json = r#"{
+                "uuid": "00c717b68e1b4213b316df82636f5e0f",
+                "private_url": "https://exercism.org/tracks/rust/exercises/poker",
+                "public_url": "https://exercism.org/tracks/rust/exercises/poker/solutions/clechasseur",
+                "status": "published",
+                "mentoring_status": "finished",
+                "published_iteration_head_tests_status": "passed",
+                "has_notifications": false,
+                "num_views": 0,
+                "num_stars": 0,
+                "num_comments": 0,
+                "num_iterations": 13,
+                "num_loc": 252,
+                "is_out_of_date": false,
+                "published_at": "2023-05-08T00:02:21Z",
+                "completed_at": "2023-05-08T00:02:21Z",
+                "updated_at": "2023-08-27T07:06:01Z",
+                "last_iterated_at": "2023-05-07T05:35:43Z",
+                "exercise": {
+                    "slug": "poker",
+                    "title": "Poker",
+                    "icon_url": "https://assets.exercism.org/exercises/poker.svg"
+                },
+                "track": {
+                    "slug": "rust",
+                    "title": "Rust",
+                    "icon_url": "https://assets.exercism.org/tracks/rust.svg"
+                }
+            }"#;
+
+            let expected = Solution {
+                uuid: "00c717b68e1b4213b316df82636f5e0f".into(),
+                private_url: "https://exercism.org/tracks/rust/exercises/poker".into(),
+                public_url:
+                    "https://exercism.org/tracks/rust/exercises/poker/solutions/clechasseur".into(),
+                status: Published,
+                mentoring_status: Finished,
+                published_iteration_head_tests_status: Passed,
+                has_notifications: false,
+                num_views: 0,
+                num_stars: 0,
+                num_comments: 0,
+                num_iterations: 13,
+                num_loc: Some(252),
+                is_out_of_date: false,
+                published_at: Some("2023-05-08T00:02:21Z".into()),
+                completed_at: Some("2023-05-08T00:02:21Z".into()),
+                updated_at: "2023-08-27T07:06:01Z".into(),
+                last_iterated_at: Some("2023-05-07T05:35:43Z".into()),
+                exercise: Exercise {
+                    name: "poker".into(),
+                    title: "Poker".into(),
+                    icon_url: "https://assets.exercism.org/exercises/poker.svg".into(),
+                },
+                track: Track {
+                    name: "rust".into(),
+                    title: "Rust".into(),
+                    icon_url: "https://assets.exercism.org/tracks/rust.svg".into(),
+                },
+            };
+            let actual: Solution = serde_json::from_str(json).unwrap();
+            assert_eq!(expected, actual);
+        }
+
+        #[test]
+        fn test_unknown() {
+            let json = r#"{
+                "uuid": "7b1fe9e73ccf44d5aa4da9b3d28da405",
+                "private_url": "https://exercism.org/tracks/rust/exercises/rlyehian",
+                "public_url": "https://exercism.org/tracks/rust/exercises/rlyehian/solutions/clechasseur",
+                "status": "syha'hnahh",
+                "mentoring_status": "or'azath",
+                "published_iteration_head_tests_status": "ah'mglw'nafh",
+                "has_notifications": false,
+                "num_views": 0,
+                "num_stars": 0,
+                "num_comments": 0,
+                "num_iterations": 13,
+                "num_loc": 252,
+                "is_out_of_date": false,
+                "published_at": "2023-05-08T00:02:21Z",
+                "completed_at": "2023-05-08T00:02:21Z",
+                "updated_at": "2023-08-27T07:06:01Z",
+                "last_iterated_at": "2023-05-07T05:35:43Z",
+                "exercise": {
+                    "slug": "rlyehian",
+                    "title": "R'lyehian",
+                    "icon_url": "https://assets.exercism.org/exercises/rlyehian.svg"
+                },
+                "track": {
+                    "slug": "rust",
+                    "title": "Rust",
+                    "icon_url": "https://assets.exercism.org/tracks/rust.svg"
+                }
+            }"#;
+
+            let expected = Solution {
+                uuid: "7b1fe9e73ccf44d5aa4da9b3d28da405".into(),
+                private_url: "https://exercism.org/tracks/rust/exercises/rlyehian".into(),
+                public_url:
+                    "https://exercism.org/tracks/rust/exercises/rlyehian/solutions/clechasseur"
+                        .into(),
+                status: Status::Unknown,
+                mentoring_status: MentoringStatus::Unknown,
+                published_iteration_head_tests_status: TestsStatus::Unknown,
+                has_notifications: false,
+                num_views: 0,
+                num_stars: 0,
+                num_comments: 0,
+                num_iterations: 13,
+                num_loc: Some(252),
+                is_out_of_date: false,
+                published_at: Some("2023-05-08T00:02:21Z".into()),
+                completed_at: Some("2023-05-08T00:02:21Z".into()),
+                updated_at: "2023-08-27T07:06:01Z".into(),
+                last_iterated_at: Some("2023-05-07T05:35:43Z".into()),
+                exercise: Exercise {
+                    name: "rlyehian".into(),
+                    title: "R'lyehian".into(),
+                    icon_url: "https://assets.exercism.org/exercises/rlyehian.svg".into(),
+                },
+                track: Track {
+                    name: "rust".into(),
+                    title: "Rust".into(),
+                    icon_url: "https://assets.exercism.org/tracks/rust.svg".into(),
+                },
+            };
+            let actual: Solution = serde_json::from_str(json).unwrap();
+            assert_eq!(expected, actual);
+        }
+    }
+}
+
+mod exercise {
+    mod deserialize {
+        use mini_exercism::api::v2::solution::Exercise;
+
+        #[test]
+        fn test_all() {
+            let json = r#"{
+                "slug": "poker",
+                "title": "Poker",
+                "icon_url": "https://assets.exercism.org/exercises/poker.svg"
+            }"#;
+
+            let expected = Exercise {
+                name: "poker".into(),
+                title: "Poker".into(),
+                icon_url: "https://assets.exercism.org/exercises/poker.svg".into(),
+            };
+            let actual: Exercise = serde_json::from_str(json).unwrap();
+            assert_eq!(expected, actual);
+        }
+    }
+}
+
+mod track {
+    mod deserialize {
+        use mini_exercism::api::v2::solution::Track;
+
+        #[test]
+        fn test_all() {
+            let json = r#"{
+                "slug": "rust",
+                "title": "Rust",
+                "icon_url": "https://assets.exercism.org/tracks/rust.svg"
+            }"#;
+
+            let expected = Track {
+                name: "rust".into(),
+                title: "Rust".into(),
+                icon_url: "https://assets.exercism.org/tracks/rust.svg".into(),
+            };
+            let actual: Track = serde_json::from_str(json).unwrap();
+            assert_eq!(expected, actual);
+        }
+    }
+}
